// src/typings/request.ts
export interface Request {
  date: string;
  time: string;
  printer: string;
  campus: string;
  document: string;
  pageRange: string;
  copies: number;
  orientation: string;
  paperSize: string;
  printSides: string;
  color: string;
  deliverDate: string;
  deliverTime: string;
  status: string;
}